Decrease build time for dbc files
