+++ selfdrive/can/dbc_template.cc
@@ -1,6 +1,4 @@
-#include <cstdint>
-
-#include "common.h"
+#include "common_dbc.h"
 
 namespace {
