+++ selfdrive/can/packer.cc
@@ -1,6 +1,4 @@
 #include <cassert>
-#include <string>
-#include <vector>
 #include <utility>
 #include <algorithm>
 #include <map>
